+++ src/library/adash-errors.ads
@@ -134,6 +134,8 @@ package Adash.Errors is
       Error_String_Index_Malformed,
       Error_Not_Taken_Apart,
       Error_No_Such_Slice,
+      Error_Needs_Bounds,
+      Error_Open_By_Element,
 
       Error_Case_Not_Discrete,
       Error_Case_Choice_Not_Static,